+++ src/MvcClient/Models/CartItem.cs
@@ -5,8 +5,8 @@ namespace MvcClient.Models
         public string Id { get; set; }
         public string ItemId { get; set; }
         public string ItemName { get; set; }
-        public decimal UnitPrice { get; set; }
-        public decimal OldUnitPrice { get; set; }
+        public double UnitPrice { get; set; }
+        public double OldUnitPrice { get; set; }
         public int Quantity { get; set; }
         public string PictureUrl { get; set; }
     }
